Fixed most of the compiler warnings and cleaned up some code in spells_effects.cpp

+++ zone/groups.cpp
@@ -1082,7 +1082,7 @@ void Client::LeaveGroup() {
 	isgrouped = false;
 }
 
-void Group::HealGroup(uint32 heal_amt, Mob* caster, int32 range)
+void Group::HealGroup(uint32 heal_amt, Mob* caster, float range)
 {
 	if (!caster)
 		return;
@@ -1120,7 +1120,7 @@ void Group::HealGroup(uint32 heal_amt, Mob* caster, int32 range)
 }
 
 
-void Group::BalanceHP(int32 penalty, int32 range, Mob* caster, int32 limit)
+void Group::BalanceHP(int32 penalty, float range, Mob* caster, int32 limit)
 {
 	if (!caster)
 		return;
@@ -1170,7 +1170,7 @@ void Group::BalanceHP(int32 penalty, int32 range, Mob* caster, int32 limit)
 	}
 }
 
-void Group::BalanceMana(int32 penalty, int32 range, Mob* caster, int32 limit)
+void Group::BalanceMana(int32 penalty, float range, Mob* caster, int32 limit)
 {
 	if (!caster)
 		return;
